data fetching environment -> interface
